Remove all optional scoped dependencies from the pom.
Gradle doesn't take care of dependencies which are scoped as 'optional',
thereby leading to transitive dependencies for projects depending on
this library.

Solution is to get rid of the dependencies scoped as 'optional' from the
dependency list, as they're anyways not supposed to be included on the
classpath or during runtime.